chore: format the code and change float to double

+++ include/falsapos.h
@@ -1,6 +1,6 @@
 #ifndef FALSAPOS_H
 #define FALSAPOS_H
 
-float falsapos(float a, float b, float e, float (*f)(float));
+double falsapos(double a, double b, double e, double (*f)(double));
 
 #endif
